Improve advanced filter and fix various issues
Fixed issue with Notes control where an error would occur if the text was changed before adding a new note.

Fixed issue with the new object control where the wrong check was being performed for a new title company.

Added the ability to filter by county name in the survey table using the advanced filter.