Clean up JobHandler and ViewPanel
Also, added a Read-Only mode and a toggle switch on the ribbon. This mode makes it so jobs opened while the switch is toggled cannot be edited, only viewed.

Still need to flesh this out some more and as of now only sets the ReadOnly flag in the JobHandler. Nothing is really using it except the JobHandler.