Fix issues with View panel null exception
Fix bug with report generation that causes an exception if the survey's description is 81 characters long.

Fix bug with JobHandler where a survey would be closed instead of opened when creating a new survey with an existing job open.

Added better check to Client.ToString() method where a more descriptive string is returned.